Please someone save me
Code for medaillons and picking it up and placing it.

It bugs when you try to pick up, place it, pick it up again, and than you can't place it anymore.

The problem lays with the counting of the children of hand_position, but I do not know how to fix this.