Medallions and torches fully working with pickup system
Read in document what I have done for this.

the order of the medallions also already matter, every holder has a specific medallion it takes in to be able to open the wall. They can of course be placed on any holder (beside torch holders), but only if all medallions are on the right medallion_holder will a print statement go off with: "open the wall!"