Mads_allemaal_gedoe_8 januari
Removed stuff from room 1

Placed all correct paths for the codeboard and gate.

Did some more stuff with the final_area like cancelling music and stopping the player from moving and making the mouse appear again.

fixed a bug with the moving wall

Fixed all the errors

Placed lights in room3

Adjusted things like anchoring and music for the final area

Solved some problem with some torches

Created an artifact with an animation at the end of the game.